+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Fixed
   * Fixed the issue with `contentWidth` as auto value.
+  
+* Changed
+  * Allow menu/sub-menu to dynamically set height based on menu items.
 
 ## 6.81.0 - (November 23, 2023)