+++ packages/terra-menu/CHANGELOG.md
@@ -6,6 +6,9 @@
   * New stylings for the focus indicator.
 * Fixed
   * Fixed the issue with `contentWidth` as auto value.
+  
+* Changed
+  * Allow menu/sub-menu to dynamically set height based on menu items.
 
 ## 6.81.0 - (November 23, 2023)